nodefs: add NewListDirStream()
